+++ app/Http/Controllers/UserController.php
@@ -6,13 +6,14 @@ use App\Handlers\ImageUploadHandler;
 use App\Model\Shop;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Mail;
 
 class UserController extends Controller{
 
     public function __construct(){
-        //权限验证
+        //登陆权限验证
         $this->middleware('auth', [
             'except' => ['index']
         ]);
@@ -20,12 +21,19 @@ class UserController extends Controller{
 
     //注册
     public function create(){
+        //用户权限验证
+        if(!Auth::user()->can('user.create')){
+            return view('error.noPermission');
+        }
         $shops=Shop::all();
         return view('user.add',['shops'=>$shops]);
     }
 
     //保存注册数据
     public function store(Request $request,ImageUploadHandler $uploader){
+        if(!Auth::user()->can('user.create')){
+            return view('error.noPermission');
+        }
         //验证数据
         $val=$this->validate($request, [
             'name' => 'required|min:3|max:20',
@@ -41,7 +49,6 @@ class UserController extends Controller{
                 'confirmed'
             ],
             'password_confirmation' => 'required|same:password',
-            'captcha' => 'required|captcha',
         ],[
             'name.required' => '用户名不能为空',
             'name.min' => '用户名不能少于三位',
@@ -54,8 +61,6 @@ class UserController extends Controller{
             'password.required' => '请输入密码',
             'password.regex' => '6-16为密码.可以是数字,字母或下划线',
             'password.confirmed' => "密码与确认密码不匹配",
-            'captcha.required' => '请输入验证码',
-            'captcha.captcha' => '验证码不正确',
             'password_confirmation.required' => "确认密码不能为空",
             'password_confirmation.same' => '',
         ]);
@@ -81,7 +86,6 @@ class UserController extends Controller{
 
     //用户首页
     public function index(){
-
         $keywords=$_GET['keywords'] ?? '';
         if($keywords!=''){
             $users=User::where('name','like','%'.$keywords.'%')
@@ -100,18 +104,23 @@ class UserController extends Controller{
 
     //修改
     public function edit(User $user){
+        if(!Auth::user()->can('user.edit')){
+            return view('error.noPermission');
+        }
         $shops=Shop::all();
         return view('user/edit',compact('user'),['shops'=>$shops]);
     }
     public function update(User $user,Request $request,ImageUploadHandler $uploader){
-        $val=$this->validate($request, [
+        if(!Auth::user()->can('user.edit')){
+            return view('error.noPermission');
+        }
+        $this->validate($request, [
             'name' => 'required|min:3|max:20',
             'email' => 'email',
             'tel' => [
                 'required',
                 'regex:/^1[3-9]\d{9}$/',
             ],
-            'captcha' => 'required|captcha'
         ],[
             'name.required' => '用户名不能为空',
             'name.min' => '用户名不能少于三位',
@@ -119,10 +128,7 @@ class UserController extends Controller{
             'email.email' => '邮箱格式不正确',
             'tel.required' => '手机号不能为空',
             'tel.regex' => '手机号格式不正确',
-            'captcha.required' => '请输入验证码',
-            'captcha.captcha' => '验证码不正确',
         ]);
-        if($val==false){ return back()->withInput();}
         //接收图片 保存图片 shopCategory文件夹名 shopcate图片文件名
         if ($request->img) {
             $result = $uploader->save($request->img, 'shopusre','user');
@@ -156,6 +162,9 @@ class UserController extends Controller{
 
     //重置密码
     public function resetPassword(User $user){
+        if(!Auth::user()->can('user.resetPassword')){
+            return view('error.noPermission');
+        }
         $string = str_random(6);
         $user->update([
             'password'=>bcrypt($string),
@@ -174,6 +183,9 @@ class UserController extends Controller{
     //删除
     public function destroy(User $user)
     {
+        if(!Auth::user()->can('user.destroy')){
+            return view('error.noPermission');
+        }
         $user->delete();
         return redirect()->route('user.index')->with('success','删除成功');
     }
